requirements.txt
Django==3.2.8
django-import-export==2.6.1
django-pandas==0.6.4
pandas==1.3.3
numpy==1.21.2
wfastcgi==3.0.0